test: remove duplicate `hot` tests
